requirements.txt
transformer
# sqlite3
requests
torch